dataproc-jobs/python/requirements.txt
google-cloud-dataproc==4.0.2
google-cloud-pubsub==2.12.0
google-cloud-storage==2.3.0